# api/swagger.yml
swagger: "2.0"
info:
  version: v1
  title: Birzzha
host: localhost:8000
basePath: /v1
schemes:
  - http
...
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  TokenHeader:
    type: apiKey
    in: header
    name: X-Token
  TokenQuery:
    type: apiKey
    in: query
    name: token
security:
  - TokenHeader: []
  - TokenQuery: []
parameters:
  RequestID:
    in: header
    name: X-Request-Id
    type: string
    format: uuid
    required: false
    description: Уникальный ID запроса. Используется для трейсинга.
paths:
  /user:
    get:
      summary: Get Current User
      operationId: getUser
      tags: [auth]
      parameters:
        - $ref: "#/parameters/RequestID"
      description: Возвращает информацию о текущем пользователе.
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/User"
        400:
          description: Bad Request
          schema:
            $ref: "#/definitions/Error"
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/Error"
...
  /topics:
    get:
      summary: Get Topics
      operationId: getTopics
      description: Получить список категорий.
      tags: [catalog]
      security: []
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: "#/definitions/Topic"
        500:
          description: Internal Server Error
...
  /tg/resolve:
    get:
      summary: Resolve Telegram Identity
      operationId: resolveTelegram
      tags: [catalog]
      description: |
        Возвращает информацию о Telegram-сущности (каналы и чат) по @username или приватной ссылке (в разработке).
        Написание ссылки или @username **не важно** (есть протокол или нет, домен, @ в начала).

        ### Пример запросов

          - https://t.me/channely_updates
          - t.me/crosser_live
          - crosser_chat
          - @zzapusk

        ### Возможные ошибки

        | Status | Code | Description |
        |:---------|:--------------|:-----------------|
        | 400 | `tg_invalid_query` | Недопустимый формат ввода |
        | 400 | `tg_entity_not_found` | Сущность не найдена |
        | 500 | `internal_error` | Внутреняя ошибка сервера |

      security: []
      parameters:
        - $ref: "#/parameters/RequestID"
        - in: query
          name: q
          type: string
          required: true
          description: запрос пользователя
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/TelegramResolveResult"
        400:
          description: Bad Request
          schema:
            $ref: "#/definitions/Error"
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/Error"
...
  /lots:
    get:
      summary: Get Lots
      operationId: getLots
      tags: [catalog]
      description: Получение списка доступных лотов.
      parameters:
        - in: query
          name: topics
          description: Фильтрация по ID категорий
          type: array
          items:
            type: integer

        - in: query
          name: members_count_from
          description: От кол-ва пдп
          type: integer

        - in: query
          name: members_count_to
          description: До кол-ва пдп
          type: integer

        - in: query
          name: price_from
          description: Цена от
          type: integer

        - in: query
          name: price_to
          description: Цена до
          type: integer

        - in: query
          name: price_per_member_from
          description: Цена пдп от
          type: number

        - in: query
          name: price_per_member_to
          description: Цена пдп до
          type: number

        - in: query
          name: daily_coverage_from
          description: суточный охват от
          type: integer

        - in: query
          name: daily_coverage_to
          description: суточный охват до
          type: integer

        - in: query
          name: price_per_view_from
          description: Цена просмотра от
          type: number

        - in: query
          name: price_per_view_to
          description: Цена просмотра до
          type: number

        - in: query
          name: monthly_income_from
          description: Месячный доход от
          type: integer

        - in: query
          name: monthly_income_to
          description: Месячный доход до
          type: integer

        - in: query
          name: payback_period_from
          description: Окупаемость в месацах от
          type: number

        - in: query
          name: payback_period_to
          description: Окупаемость в месацах до
          type: number

        - in: query
          name: sort_by
          type: string
          enum:
            - members_count
            - price
            - price_per_member
            - price_per_view
            - daily_coverage
            - monthly_income
            - payback_period
            - created_at
          default: created_at

        - in: query
          name: sort_by_type
          type: string
          enum:
            - asc
            - desc
          default: desc
          description: сортировать от меньшего к большему (`asc`), или от большего к меньшему (`desc`)

        - in: query
          name: offset
          type: integer
          description: сдвиг начала каждой страницы на указанное количество единиц

        - in: query
          name: limit
          type: integer
          description: максимальное количество объектов, которое может вернуть запрос

      security:
        - {}
        - TokenHeader: []
        - TokenQuery: []
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/LotList"
        400:
          description: Bad Request
          schema:
            $ref: "#/definitions/Error"
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/Error"
...
  /lots/{id}/favorite:
    parameters:
      - name: id
        in: path
        type: integer
        required: true
        description: ID лота
    post:
      summary: Change Favorite Status
      description: Изменение статуса избранности лота.
      operationId: toggleLotFavorite
      tags: [catalog]
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/LotFavoriteStatus"
        400:
          description: Bad Request
          schema:
            $ref: "#/definitions/Error"
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/Error"
  /lots/{id}/similars:
    parameters:
      - name: id
        in: path
        type: integer
        required: true
        description: ID лота
      - in: query
        name: offset
        type: integer
        required: false
        description: Офсет похожих лотов.
        default: 0
      - in: query
        name: limit
        type: integer
        required: false
        description: Лимит похожих лотов.
        default: 10
    get:
      summary: Get Similar Lots
      description: Получить похожие лоты.
      operationId: getSimilarLots
      tags: [catalog]
      security: []
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/LotList"
        400:
          description: Bad Request
          schema:
            $ref: "#/definitions/Error"
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/Error"
  /lots/filter-boundaries:
    get:
      summary: Get Filter Boundaries
      description: Получить границы фильтрации
      operationId: getFilterBoundaries
      security: []
      parameters:
        - name: topics
          in: query
          type: array
          description: ID категорий для фильтрации границ
          items:
            type: integer

      tags: [catalog]
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/LotFilterBoundaries"
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/Error"
...
definitions:
...
  LotFilterBoundaries:
    type: object
    required:
      - members_count
      - price
      - price_per_member
      - price_per_view
      - daily_coverage
      - monthly_income
      - payback_period
    properties:
      members_count:
        $ref: "#/definitions/IntBoundaries"
      price:
        $ref: "#/definitions/IntBoundaries"
      price_per_member:
        $ref: "#/definitions/FloatBoundaries"
      price_per_view:
        $ref: "#/definitions/FloatBoundaries"
      daily_coverage:
        $ref: "#/definitions/IntBoundaries"
      monthly_income:
        $ref: "#/definitions/IntBoundaries"
      payback_period:
        $ref: "#/definitions/FloatBoundaries"
  FloatBoundaries:
    type: object
    required:
      - from
      - to
    properties:
      from:
        type: number
        description: от
      to:
        type: number
        description: to
  IntBoundaries:
    type: object
    required:
      - from
      - to
    properties:
      from:
        type: integer
        description: от
      to:
        type: integer
        description: to
  LotList:
    type: object
    description: Список лотов
    required:
      - total
      - items
    properties:
      total:
        type: integer
        description: кол-во объектов подходщих под фильтр
      items:
        type: array
        items:
          $ref: "#/definitions/LotListItem"
  LotListItem:
    type: object
    description: Элемент списка лотов
    required:
      - id
      - name
      - avatar
      - link
      - username
      - price
      - comment
      - metrics
      - in_favorites
      - created_at
    properties:
      id:
        type: integer
        description: ID лота
      name:
        type: string
        description: Название лота (канала) в Telegram
      avatar:
        type: string
        format: url
        description: Аватарка лота
      username:
        type: string
        description: "@username канала (может быть null)"
      link:
        type: string
        description: Ссылка для вступления (как приватная так и публичная)
        format: url
      price:
        $ref: "#/definitions/LotPrice"
      comment:
        type: string
        description: Комментарий к лоту
      metrics:
        $ref: "#/definitions/LotMetrics"
      topics:
        type: array
        items:
          type: integer
          minLength: 1
          maxLength: 3
      in_favorites:
        type: boolean
        description: True, если лот в избранном
      created_at:
        type: integer
        description: Дата создания
...
  LotMetrics:
    type: object
    required:
      - members_count
      - daily_coverage
      - monthly_income
      - price_per_member
      - price_per_view
      - payback_period
    properties:
      members_count:
        type: integer
        description: Кол-во участников в канале
      daily_coverage:
        type: integer
        description: Дневной охват
      monthly_income:
        type: integer
        description: Прибыль в месяц
      price_per_member:
        type: number
        description: Цена за подписчика
      price_per_view:
        type: number
        description: Цена за просмотр
      payback_period:
        type: number
        description: Окупаемость
  LotPrice:
    type: object
    required:
      - current
      - previous
      - is_bargain
    properties:
      current:
        type: integer
        description: Цена канала
      previous:
        type: integer
        description: Предыдущая цена канала
      is_bargain:
        type: boolean
        description: Возможен ли торг
...
  TelegramResolveResult:
    type: object
    description: Сущность соотвествующая запросу.
    properties:
      channel:
        type: object
        description: Канал
        required:
          - id
          - name
          - avatar
          - members_count
          - username
          - daily_average
          - description
        properties:
          id:
            type: integer
            description: Уникальный ID канала в Telegramа
          name:
            type: string
            description: Название канала
          avatar:
            type: string
            description: URL аватарки канала (может быть null)
          description:
            type: string
            description: Описание
          members_count:
            type: integer
            description: Кол-во участников в канале
          username:
            type: string
            description: username канала (может быть null)
          daily_average:
            type: integer
            description: Среднее кол-во просмотров в день
      group:
        type: object
        description: Групповой чат
        required:
          - id
          - name
          - avatar
          - members_count
          - username
          - daily_average
          - description
        properties:
          id:
            type: integer
            description: Уникальный ID чата в Telegrama
          name:
            type: string
            description: Название чата
          description:
            type: string
            description: Описание
          avatar:
            type: string
            description: URL аватарки чата (может быть null)
          members_count:
            type: integer
            description: Кол-во участников в чате
          username:
            type: string
            description: username чата (может быть null)
  Topic:
    type: object
    description: Категория лота
    required:
      - id
      - name
      - slug
      - created_at
    properties:
      id:
        type: integer
        description: Уникальный ID категории.
      name:
        type: string
        description: Название категории на русском.
      slug:
        type: string
        description: Slug (уникальный)
      created_at:
        type: integer
        description: Дата создания
  User:
    type: object
    description: Объект пользователя
    required:
      - id
      - telegram
      - first_name
      - last_name
      - avatar
      - is_admin
      - joined_at
    properties:
      id:
        type: integer
        description: Уникальный ID пользователя в Birzzha.
      telegram:
        type: object
        description: Информация о пользователе из Telegram
        required:
          - id
          - username
        properties:
          id:
            type: integer
            description: ID пользователя в Telegram
          username:
            type: string
            description: Username пользователя в Telegram
      first_name:
        type: string
        description: Имя пользователя в Telegram
      last_name:
        type: string
        description: Фамилия пользователя в Telegram (может быть `null`)
      avatar:
        type: string
        description: Path to avatar
      is_admin:
        type: boolean
        description: True, если пользователь админ Birzzha.
      joined_at:
        type: integer
        description: Дата и время регистрации на бирже, в Unix-time.
  Error:
    type: object
    required:
      - code
      - description
    properties:
      code:
        type: string
        enum: [test]
        description: Код ошибки
      description:
        type: string
        description: Описание ошибки
...
  LotFavoriteStatus:
    type: object
    description: Ответ после смены статуса избранности лота.
    required:
      - in_favorites
    properties:
      in_favorites:
        type: boolean
        description: Статус избранности лота.
...